+++ extensions/measurement-tracking/src/panels/PanelStudyBrowserTracking/PanelStudyBrowserTracking.tsx
@@ -95,7 +95,7 @@ export default function PanelStudyBrowserTracking({
           seriesNumber: ds.SeriesNumber,
           modality: ds.Modality,
           seriesDate: ds.SeriesDate ? new Date(ds.SeriesDate).toLocaleDateString() : '',
-          numInstances: ds.numImageFrames,
+          numInstances: ds.numImageFrames ?? ds.instances?.length,
           loadingProgress,
           countIcon: ds.countIcon,
           messages: ds.messages,