+++ tests/RTDataOverlayForUnreferencedDisplaySetNoHydration.spec.ts
@@ -1,5 +1,4 @@
-import { test } from 'playwright-test-coverage';
-import { visitStudy, checkForScreenshot, screenShotPaths } from './utils';
+import { checkForScreenshot, screenShotPaths, test, visitStudy } from './utils';
 import { press } from './utils/keyboardUtils';
 import { assertNumberOfModalityLoadBadges } from './utils/assertions';
 
@@ -11,17 +10,17 @@ test.beforeEach(async ({ page }) => {
 
 test('should overlay an unhydrated RTSTRUCT over a display set that the RTSTRUCT does NOT reference', async ({
   page,
+  viewportPageObject,
 }) => {
-  await page.getByTestId('dataOverlayMenu-default-btn').click();
-  await page.getByTestId('AddSegmentationDataOverlay-default').click();
-  await page.getByText('SELECT A SEGMENTATION').click();
-  await page.getByTestId('Contours on PET').click();
+  const dataOverlayPageObject = viewportPageObject.getById('default').overlayMenu.dataOverlay;
+  await dataOverlayPageObject.toggle();
+  await dataOverlayPageObject.addSegmentation('Contours on PET');
 
   // Adding an overlay should not show the LOAD button.
   assertNumberOfModalityLoadBadges({ page, expectedCount: 0 });
 
   // Hide the overlay menu.
-  await page.getByTestId('dataOverlayMenu-default-btn').click();
+  await dataOverlayPageObject.toggle();
 
   await page.waitForTimeout(5000);