+++ tests/SEGHydrationFrom3DFourUp.spec.ts
@@ -1,10 +1,10 @@
-import { test } from 'playwright-test-coverage';
 import {
-  visitStudy,
-  checkForScreenshot,
-  screenShotPaths,
-  reduce3DViewportSize,
   attemptAction,
+  checkForScreenshot,
+  reduce3DViewportSize,
+  screenShotPaths,
+  test,
+  visitStudy,
 } from './utils';
 
 test.beforeEach(async ({ page }) => {
@@ -14,9 +14,13 @@ test.beforeEach(async ({ page }) => {
 });
 
 test.describe('3D four up SEG hydration', async () => {
-  test('should properly hydrate SEG from 3D four up layout', async ({ page }) => {
-    await page.getByTestId('Layout').click();
-    await page.getByTestId('3D four up').click();
+  test('should properly hydrate SEG from 3D four up layout', async ({
+    page,
+    DOMOverlayPageObject,
+    leftPanelPageObject,
+    mainToolbarPageObject,
+  }) => {
+    await mainToolbarPageObject.layoutSelection.threeDFourUp.click();
 
     await attemptAction(() => reduce3DViewportSize(page), 10, 100);
 
@@ -28,7 +32,7 @@ test.describe('3D four up SEG hydration', async () => {
       screenShotPaths.segHydrationFrom3DFourUp.threeDFourUpBeforeSEG
     );
 
-    await page.getByTestId('study-browser-thumbnail-no-image').dblclick();
+    await leftPanelPageObject.loadSeriesByDescription('SEG');
 
     await page.waitForTimeout(5000);
     await checkForScreenshot(
@@ -37,7 +41,7 @@ test.describe('3D four up SEG hydration', async () => {
       screenShotPaths.segHydrationFrom3DFourUp.threeDFourUpAfterSEG
     );
 
-    await page.getByTestId('yes-hydrate-btn').click();
+    await DOMOverlayPageObject.viewport.segmentationHydration.yes.click();
 
     await page.waitForTimeout(5000);
     await checkForScreenshot(